refactored business logic out of controller and into a service class. Controller integration test with mockmvc
