+++ src/main/java/com/example/cicdtodoapp/controllers/TaskController.java
@@ -1,21 +1,26 @@
 package com.example.cicdtodoapp.controllers;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import com.example.cicdtodoapp.models.Task;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
+import com.example.cicdtodoapp.models.Task;
+import com.example.cicdtodoapp.services.TaskService;
+
+@RestController
+@RequestMapping("/api/v1/tasks")
 public class TaskController {
 
-	private List<Task> tasks = new ArrayList<>();
+	private TaskService taskService = new TaskService(); // not @Service yet
+	
+	@GetMapping
+	public List<Task> getTasks() {
+		return taskService.getTasks();
+	}
 	
 	public void addTask(Task task) {
-		tasks.add(task);
-		
+		taskService.addTask(task);
 	}
-
-	public List<Task> getTasks() {
-		return tasks;
-	}
-
 }
